master/requirements.txt
# Master Flask Dependencies
Flask>=2.3.0
Flask-SocketIO>=5.3.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0
python-socketio>=5.8.0
eventlet>=0.33.0

# Camera (Pi Camera v2) - CCTV Only
picamera>=1.13
picamera2>=0.3
opencv-python>=4.8.0
numpy>=1.24.0

# Hardware Interface (Raspberry Pi GPIO + SPI)
pyserial>=3.5
RPi.GPIO>=0.7.0
spidev>=3.5

# Line Following (IR Sensors) - Physical pins defined in R